requirements.txt
click
rich
pyOpenSSL
semgrep
rsa
pyyaml
jinja2
weasyprint
oqs
astunparse
flask
tqdm
pytest
requests
